1.修改播放结束卡顿bug ： 原因播放结束算法会调用pause()进行暂停，PAUSE()函数设计有bug,会重新计算前面所有事件的时间  值，大量的事件计算影响了发音的流畅
2.修改chorus一处注释bug

+++ Tau/src/core/Synth/Synther/Synther.h
@@ -420,7 +420,7 @@ namespace tau
 		//不受真实时间流逝影响，比如已处理了2000个点采样点，每个采样点的采样时间为1.0f/44100.0f
 		//则当前sec = 2000 * 1.0f/44100.0f = 0.045s = 45ms,
 		//即表示目前播放时间已流逝0.045s
-		float sec = 0;
+		double sec = 0;
 
 		//所有正在发声的区域
 		RegionSounder* totalRegionSounders[100000] = { nullptr };
